backend/requirements.txt
# MAESTRO Backend Requirements
# Install: pip install -r requirements.txt

# ── Web Framework ─────────────────────────────────────────────
fastapi>=0.109.0
uvicorn[standard]>=0.25.0
websockets>=12.0
python-multipart>=0.0.6
aiofiles>=23.2.1

# ── Audio Capture ─────────────────────────────────────────────
sounddevice>=0.4.6
# numpy and scipy are provided by system packages in EndeavourOS/Arch
numpy>=1.26.3
scipy>=1.12.0

# ── AI Models ─────────────────────────────────────────────────
# Version 2.10.0+ is required for Python 3.14 compatibility
torch>=2.10.0
torchaudio>=2.10.0
faster-whisper>=0.10.0
onnxruntime>=1.16.3
transformers>=4.37.0
sentence-transformers>=2.3.1
huggingface-hub>=0.20.3

# ── Agentic Framework ─────────────────────────────────────────
langchain>=0.1.0
langchain-google-genai>=0.0.6
langgraph>=0.0.26
google-generativeai>=0.3.2

# ── Memory & State ────────────────────────────────────────────
chromadb>=0.4.22
redis>=5.0.1

# ── Utilities ─────────────────────────────────────────────────
python-dotenv>=1.0.0
pyyaml>=6.0.1
httpx>=0.26.0
slack-sdk>=3.26.2
